+++ compiler/src/regalloc.ml
@@ -722,6 +722,15 @@ let two_phase_coloring
         A.set i y a
     ) schedule
 
+let check_allocated
+      (vars: (int, var list) Hashtbl.t)
+      (a: A.allocation) : unit =
+  match Hashtbl.fold (fun i x m -> if A.mem i a then m else x @ m) vars [] with
+  | [] -> ()
+  | m ->
+     hierror_reg ~loc:Lnone "variables { %a } remain unallocated"
+       (pp_list "; " (Printer.pp_var ~debug:true)) m
+
 let greedy_allocation
     (vars: int Hv.t)
     (nv: int) (cnf: conflicts)
@@ -729,6 +738,7 @@ let greedy_allocation
     (a: A.allocation) : unit =
   let scalars : (int, var list) Hashtbl.t = Hashtbl.create nv in
   let vectors : (int, var list) Hashtbl.t = Hashtbl.create nv in
+  let flags : (int, var list) Hashtbl.t = Hashtbl.create nv in
   let push_var tbl i v =
     match Hashtbl.find tbl i with
     | old -> Hashtbl.replace tbl i (v :: old)
@@ -738,11 +748,12 @@ let greedy_allocation
       match kind_of_type v.v_ty with
       | Word -> push_var scalars i v
       | Vector -> push_var vectors i v
-      | Flag
+      | Flag -> push_var flags i v
       | Unknown _ -> ()
       ) vars;
   two_phase_coloring X64.allocatable scalars cnf fr a;
   two_phase_coloring X64.xmm_allocatable vectors cnf fr a;
+  check_allocated flags a;
   ()
 
 let var_subst_of_allocation (vars: int Hv.t)